requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
APScheduler==3.11.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
discord.py==2.6.4
dnspython==2.8.0
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
motor==3.7.1
multidict==6.7.0
nextcord==3.1.1
propcache==0.4.1
py-cord==2.6.1
pycparser==2.23
pymongo==4.15.3
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
yarl==1.22.0
